requirements.txt
python-telegram-bot
requests
cohere


